requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.1
Django==4.2.2
django-celery-beat==2.5.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docopt==0.6.2
drf-yasg==1.21.7
eventlet==0.33.3
greenlet==2.0.2
idna==3.4
inflection==0.5.1
kombu==5.3.1
msal==1.22.0
packaging==23.1
pipreqs==0.4.13
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.7.0
pyTelegramBotAPI==4.12.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
telebot==0.0.5
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
yarg==0.1.9
zope.event==5.0
zope.interface==6.0
